third_party/WebKit/LayoutTests/inspector/tracing/scroll-invalidations.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script src="../../http/tests/inspector/timeline-test.js"></script>
<script>
function scrollAndDisplay(callback)
{
    scrollTo(0, 200);
    if (window.testRunner)
        testRunner.layoutAndPaintAsyncThen(callback);
}

function test()
{
    var currentPanel = WebInspector.inspectorView.currentPanel();
    InspectorTest.assertEquals(currentPanel._panelName, "timeline", "Current panel should be the timeline.");
    Runtime.experiments.enableForTest("timelineInvalidationTracking");

    InspectorTest.invokeAsyncWithTimeline("scrollAndDisplay", onRecordingDone);

    function onRecordingDone()
    {
        var record = InspectorTest.findFirstTimelineRecord(WebInspector.TimelineModel.RecordType.Paint);
        InspectorTest.addArray(record._event.invalidationTrackingEvents, InspectorTest.InvalidationFormatters, "", "Scroll invalidations");
        InspectorTest.completeTest();
    };
}
</script>
</head>
<body onload="runTest()">
<p>Tests invalidations produced by scrolling a page with position: fixed elements.</p>
<div style="width: 400px; height: 2000px; background-color: grey"></div>
<div style="position: fixed; left: 50px; top: 100px; width: 50px; height: 50px; background-color: rgba(255, 100, 100, 0.6)"></div>
</body>
</html>
