third_party/WebKit/LayoutTests/platform/android/inspector/tracing/scroll-invalidations-expected.txt

Tests invalidations produced by scrolling a page with position: fixed elements.

Scroll invalidations[
    {
        cause : {reason: Scroll with viewport-constrained element, stackTrace: .../inspector/tracing/scroll-invalidations.html:11}
        changedAttribute : undefined
        changedClass : undefined
        changedId : undefined
        changedPseudo : undefined
        extraData : undefined
        nodeName : "DIV"
        selectorPart : undefined
        type : "ScrollInvalidationTracking"
    }
]
